Typo:  "TWELV" --> "TWELVE"
Added note at top, since LaTeX style writers like to force name
changes when someone else makes a change to the style.  I'm sending a
note to the author reporting the error and asking for an updated
version.

This cannot be distributed in this form.  This will be resolved when I
get a response from Ulf Lindgren.

Typo reported by rwolff@noao.edu (Richard Wolff).