When we know the compiler is GCC, always add the -Wall and -Wstrict-prototypes options.  This will make it a lot easier to keep warnings under control in the first place in the future.
There is one known warning at this time, caught by the -Wstrict-prototypes
option.  In Modules/main.c, the declaration of getopt() without parameters
gets a complaint (rightly) that it is not a proper prototype.  The lack of
a complete prototype information should be corrected when the right
portability conditions have been identified.

Approved by the Guido.