d'oh! could have _sworn_ I checked this in last weekend.
Make dumbdbm merely "dumb", rather than "terminally broken". Without this
patch, it's almost impossible to use dumbdbm _without_ causing horrible
datalossage. With this patch, dumbdbm passes my own horrible torture test,
as well as the roundup test suite.

dumbdbm really could do with a smidgin of a rewrite or two, but that's not
suitable for the release21-maint branch.

This patch should go into the trunk as well.