long_format(), long_lshift():  Someone on c.l.py is trying to boost SHIFT and MASK, and widen digit.  One problem is that code of the form
    digit << small_integer

implicitly assumes that the result fits in an int or unsigned int
(platform-dependent, but "int sized" in any case), since digit is
promoted "just" to int or unsigned via the usual integer promotions.
But if digit is typedef'ed as unsigned int, this loses information.
The cure for this is just to cast digit to twodigits first.