morphos: fix build_scripts phase during installation
idle3 was removed from scripts to install if platform is morphos-ppc.
But the path of idle3 was replaced by an empty string in this case.
This empty string is transformed as the temporary build path,
taken in account as a "script" (and not a directory) and renamed
by setup.py causing rename operation for furthers scripts to fail.
So idle3 is re-enable (no empty strings), but it will be removed
during release target call.